// SGUEES-SPA/src/app/pages/Accounting/con-parametro/models/con-parametro.ts
export interface ConParametro {
	CORR_EMPRESA: number;
	NOMBRE_PUESTO1: string;
	DESCRIPCION_PUESTO1: string;
	NOMBRE_PUESTO2: string;
	DESCRIPCION_PUESTO2: string;
	NOMBRE_PUESTO3: string;
	DESCRIPCION_PUESTO3: string;
	NIVEL_CUENTA_MAYOR: number;
	CORR_CENTRO_COSTO_DEF: number;
	CORR_MONEDA: number;
	CUENTA_CONTABLE_PERDIDA: string;
	CUENTA_CONTABLE_GANANCIA: string;
	CUENTA_CONTABLE_IVA_DEBITO: string;
	CUENTA_CONTABLE_IVA_CREDITO: string;
	CUENTA_CONTABLE_IVA_RETENIDO: string;
	CUENTA_CONTABLE_IVA_PERCIBIDO: string;
	CUENTA_CONTABLE_RENTA: string;
	CUENTA_CONTABLE_CAJA: string;
	CUENTA_CONTABLE_CAJA_CHICA: string;
	APLICAR_DOC_CONTA: boolean;
	PERIODO_MOSTRAR: string;
	CARACTER_SEPARADOR: string;
	CUENTA_CONTABLE_CAMBIO_DIFERENCIAL: string;
	CORR_CLASE_PARTIDA_DEFAULT: number;
	OCULTA_CLASE_PARTIDA_BANCOS: boolean;
	INCLUIR_PARTIDA_LIQUIDACION: boolean;
	INCLUIR_PARTIDA_CIERRE: boolean;
	USA_AUXILIARES: boolean;
	MOSTRAR_FECHA_IMPRESION: boolean;
	PREFIJO: string;
	CUENTA_CONTABLE_CAMBIO_DIF_GASTO: string;
}